general_text_tools: Toggle CamelCase/underscores_style
Useful tool when developing Sublime Text plugins - Quickly toggle between
CamelCase and underscores_style: the two most popular styles for naming
variables and other things.

Note: A little limitation is that the conversion to CamelCase is a bit lossy,
in that the numbers always end up being left-associative now. Perhaps this is
something another option could sort out, but that's getting into the territory'
of arguments overkill...